Parking creation

+++ src/main/java/pl/dmcs/nsai/struts2/repositories/ParkingDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import pl.dmcs.nsai.struts2.entities.ParkingData;
 
-public interface ParkingDAO extends CrudRepository<ParkingData, Long> {
+public interface ParkingDAO extends CrudRepository<ParkingData, Integer> {
 	
 }
